Switch define and create commands to new infrastructure.
